backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
pymupdf
httpx
langfuse
supabase
python-dotenv
nltk
